+++ federation/utils/text.py
@@ -3,3 +3,10 @@ def decode_if_bytes(text):
         return text.decode("utf-8")
     except AttributeError:
         return text
+
+
+def encode_if_text(text):
+    try:
+        return bytes(text, encoding="utf-8")
+    except TypeError:
+        return text